LDLS/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=0.11.0=pypi_0
alabaster=0.7.12=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
attrs=20.3.0=pypi_0
babel=2.9.0=pypi_0
backcall=0.2.0=pypi_0
binutils_impl_linux-64=2.35.1=h17ad2fc_0
binutils_linux-64=2.35=hc3fd857_29
bleach=1.5.0=pypi_0
c-ares=1.16.1=h516909a_3
ca-certificates=2020.11.8=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=4.1.1=pypi_0
certifi=2020.11.8=py36h5fab9bb_0
cffi=1.14.3=pypi_0
chardet=3.0.4=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cupy=8.1.0=py36hc4700d2_0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
docutils=0.16=pypi_0
entrypoints=0.3=pypi_0
fastrlock=0.5=py36h831f99a_1
freetype=2.10.4=h7ca028e_0
gast=0.3.3=pypi_0
gcc_impl_linux-64=9.3.0=h28f5a38_17
gcc_linux-64=9.3.0=h7247604_29
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.33.2=py36he0f7d3b_2
gxx_impl_linux-64=9.3.0=h53cdd4c_17
gxx_linux-64=9.3.0=h0d07fa4_29
h5py=2.10.0=pypi_0
hdf5=1.10.6=nompi_h1022a3e_1110
html5lib=0.9999999=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
imagesize=1.2.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=2.0.0=py_1
ipdb=0.13.4=pypi_0
ipykernel=5.3.4=pypi_0
ipyparallel=6.3.0=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=pypi_0
jpeg=9d=h36c2ea0_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pypi_0
kernel-headers_linux-64=2.6.32=h77966d4_13
kiwisolver=1.3.1=pypi_0
krb5=1.17.1=hfafb76e_3
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35.1=hed1e6ac_0
libblas=3.9.0=2_openblas
libcblas=3.9.0=2_openblas
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.2.1=he1b5a44_1007
libgcc-devel_linux-64=9.3.0=hfd08b2a_17
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libgomp=9.3.0=h5dbcf3e_17
libgpuarray=0.7.6=h14c3975_1003
liblapack=3.9.0=2_openblas
libnghttp2=1.41.0=h8cfc5f6_2
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.13.0.1=h8b12597_0
libssh2=1.9.0=hab1572f_5
libstdcxx-devel_linux-64=9.3.0=h4084dd6_17
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h4f3a223_6
libwebp-base=1.1.0=h36c2ea0_3
llvm-openmp=11.0.0=hfc4b9b4_1
llvmlite=0.34.0=pypi_0
lz4-c=1.9.2=he1b5a44_3
mako=1.1.3=pyh9f0ad1d_0
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py36he6145b8_2
matplotlib=3.3.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.4=h726a3e6_304
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.0.8=pypi_0
nccl=2.7.8.1=hd6f8bf8_1
ncurses=6.2=h58526e2_3
nest-asyncio=1.4.3=pypi_0
networkx=2.5=pypi_0
ninja=1.10.1=hfc4b9b4_2
nose=1.3.7=pypi_0
notebook=6.1.5=pypi_0
numba=0.51.2=pypi_0
numpy=1.19.4=py36h8732dcd_1
oauthlib=3.1.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.4.0.46=pypi_0
openssl=1.1.1h=h516909a_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.7.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.1=py36h10ecd5c_0
pip=20.2.4=py_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.8=pypi_0
protobuf=3.14.0=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.2=pypi_0
pygpu=0.7.6=py36h68bb277_1002
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
python=3.6.11=h4d41432_2_cpython
python-dateutil=2.8.1=pypi_0
python_abi=3.6=1_cp36m
pytorch=1.4.0=py3.6_cuda10.0.130_cudnn7.6.3_0
pytz=2020.4=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=pypi_0
qtconsole=4.7.7=pypi_0
qtpy=1.9.0=pypi_0
readline=8.0=he28a2e2_2
requests=2.25.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py36h9880bd3_2
shapely=1.7.1=pypi_0
six=1.15.0=pyh9f0ad1d_0
snowballstemmer=2.0.0=pypi_0
sphinx=3.3.1=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=1.0.3=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.4=pypi_0
sqlite=3.33.0=h4cf870e_1
sysroot_linux-64=2.12=h77966d4_13
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=py_0
tensorflow=1.3.0=pypi_0
tensorflow-base=1.14.0=py36hc3e5e64_0
tensorflow-estimator=2.3.0=pypi_0
tensorflow-tensorboard=0.1.8=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
theano=1.0.4=py36hf484d3e_1000
threadpoolctl=2.1.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.10=hed695b0_1
torchvision=0.5.0=py36_cu100
tornado=6.1=pypi_0
traitlets=4.3.3=pypi_0
urllib3=1.26.2=pypi_0
v=0.0.0=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.35.1=pyh9f0ad1d_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=py36h8c4c3a4_1
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.5=h6597ccf_2
